libwasi: move some stuff to wasi_path.[ch]
